Revert "feat(route): 公視新聞網 (#10236)"
This reverts commit fc34da7d6f9ab8484faa42ae7a8ae2e459fe64d4.
