+++ lib/v2/pts/dailynews.js
@@ -1,0 +1,52 @@
+const got = require('@/utils/got');
+const cheerio = require('cheerio');
+const timezone = require('@/utils/timezone');
+const { parseDate } = require('@/utils/parse-date');
+
+module.exports = async (ctx) => {
+    const rootUrl = 'https://news.pts.org.tw';
+    const currentUrl = `${rootUrl}/dailynews`;
+
+    const response = await got({
+        method: 'get',
+        url: currentUrl,
+    });
+
+    const $ = cheerio.load(response.data);
+
+    const list = $('h2 a')
+        .map((_, item) => {
+            item = $(item);
+
+            return {
+                title: item.text(),
+                link: item.attr('href'),
+            };
+        })
+        .get();
+
+    const items = await Promise.all(
+        list.map((item) =>
+            ctx.cache.tryGet(item.link, async () => {
+                const detailResponse = await got({
+                    method: 'get',
+                    url: item.link,
+                });
+
+                const content = cheerio.load(detailResponse.data);
+
+                item.author = content('meta[property="article:author"]').attr('content');
+                item.pubDate = timezone(parseDate(content('meta[property="article:published_time"]').attr('content')), +8);
+                item.description = `<img src="${content('meta[property="og:image"]').attr('content')}">${content('.post-article').html()}`;
+
+                return item;
+            })
+        )
+    );
+
+    ctx.state.data = {
+        title: '即時 ｜ 公視新聞網 PNN',
+        link: currentUrl,
+        item: items,
+    };
+};
