+++ lib/v2/pts/maintainer.js
@@ -1,9 +1,3 @@
 module.exports = {
-    '/category/:id': ['nczitzk'],
-    '/curations': ['nczitzk'],
     '/dailynews': ['nczitzk'],
-    '/opinion': ['nczitzk'],
-    '/projects': ['nczitzk'],
-    '/report': ['nczitzk'],
-    '/tag/:id': ['nczitzk'],
 };
