+++ lib/v2/pts/radar.js
@@ -3,46 +3,10 @@ module.exports = {
         _name: '公視新聞網 PNN',
         news: [
             {
-                title: '即時新聞',
+                title: '即時',
                 docs: 'https://docs.rsshub.app/traditional-media.html#gong-shi-xin-wen-wang-ji-shi-xin-wen',
-                source: ['/dailynews', '/'],
+                source: ['/dailynews'],
                 target: '/pts/dailynews',
-            },
-            {
-                title: '專題策展',
-                docs: 'https://docs.rsshub.app/traditional-media.html#gong-shi-xin-wen-wang-zhuan-ti-ce-zhan',
-                source: ['/curations', '/'],
-                target: '/pts/curations',
-            },
-            {
-                title: '觀點',
-                docs: 'https://docs.rsshub.app/traditional-media.html#gong-shi-xin-wen-wang-guan-dian',
-                source: ['/opinion', '/'],
-                target: '/pts/opinion',
-            },
-            {
-                title: '數位敘事',
-                docs: 'https://docs.rsshub.app/traditional-media.html#gong-shi-xin-wen-wang-shu-wei-xu-shi',
-                source: ['/projects', '/'],
-                target: '/pts/projects',
-            },
-            {
-                title: '深度報導',
-                docs: 'https://docs.rsshub.app/traditional-media.html#gong-shi-xin-wen-wang-shen-du-bao-dao',
-                source: ['/report', '/'],
-                target: '/pts/report',
-            },
-            {
-                title: '分類',
-                docs: 'https://docs.rsshub.app/traditional-media.html#gong-shi-xin-wen-wang-fen-lei',
-                source: ['/category/:id', '/'],
-                target: '/pts/category/:id',
-            },
-            {
-                title: '標籤',
-                docs: 'https://docs.rsshub.app/traditional-media.html#gong-shi-xin-wen-wang-biao-qian',
-                source: ['/tag/:id', '/'],
-                target: '/pts/tag/:id',
             },
         ],
     },
