+++ lib/v2/pts/router.js
@@ -1,6 +1,3 @@
 module.exports = function (router) {
-    router.get('/curations', require('./curations'));
-    router.get('/projects', require('./projects'));
-    router.get(/([\w-/]+)?/, require('./index'));
-    router.get('', require('./index'));
+    router.get('/dailynews', require('./dailynews'));
 };
