--- lib/v2/pts/templates/description.art
@@ -1,6 +1,0 @@
-{{ if image }}
-<img src="{{ image }}">
-{{ /if }}
-{{ if description }}
-{{@ description }}
-{{ /if }}
